+++ database/migrations/2026_04_13_173221_add_columns_to_log_aset_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('log_aset', function (Blueprint $table) {
-            $table->unsignedInteger('lokasi_id')->nullable()->after('keterangan');
+            $table->unsignedBigInteger('lokasi_id')->nullable()->after('keterangan');
             $table->unsignedInteger('id_divisi')->nullable()->after('lokasi_id');
             $table->string('foto_bukti')->nullable()->after('id_divisi');
